+++ packages/oauth/oauth-provider/src/router/create-oauth-middleware.ts
@@ -168,8 +168,14 @@ export function createOAuthMiddleware<
         .parseAsync(payload, { path: ['body'] })
         .catch(throwInvalidRequest)
 
+      const dpopProof = await server.checkDpopProof(
+        req.method!,
+        this.url,
+        req.headers,
+      )
+
       try {
-        await server.revoke(credentials, tokenIdentification)
+        await server.revoke(credentials, tokenIdentification, dpopProof)
       } catch (err) {
         // > Note: invalid tokens do not cause an error response since the
         // > client cannot handle such an error in a reasonable way.  Moreover,