+++ packages/pds/src/account-manager/oauth-store.ts
@@ -428,9 +428,9 @@ export class OAuthStore
     await this.db.executeWithRetry(authRequestHelper.removeByIdQB(this.db, id))
   }
 
-  async findRequestByCode(code: Code): Promise<FoundRequestResult | null> {
+  async consumeRequestCode(code: Code): Promise<FoundRequestResult | null> {
     const row = await authRequestHelper
-      .findByCodeQB(this.db, code)
+      .consumeByCodeQB(this.db, code)
       .executeTakeFirst()
     return row ? authRequestHelper.rowToFoundRequestResult(row) : null
   }