+++ packages/pds/src/config/config.ts
@@ -226,7 +226,6 @@ export const envToCfg = (env: ServerEnvironment): ServerConfig => {
   const rateLimitsCfg: ServerConfig['rateLimits'] = env.rateLimitsEnabled
     ? {
         enabled: true,
-        mode: redisCfg !== null ? 'redis' : 'memory',
         bypassKey: env.rateLimitBypassKey,
         bypassIps: env.rateLimitBypassIps?.map((ipOrCidr) =>
           ipOrCidr.split('/')[0]?.trim(),
@@ -491,7 +490,6 @@ export type RedisScratchConfig = {
 export type RateLimitsConfig =
   | {
       enabled: true
-      mode: 'memory' | 'redis'
       bypassKey?: string
       bypassIps?: string[]
     }